\name{scottish.lip.cancer}
\alias{scottish.lip.cancer}
\title{Cases of Lip Cancer in Scottish Counties}
\description{The number of cases of lip cancer registered
  in each of the 56 (pre-reoreganisation) counties of
  Scotland, together with  the expected number of cases given
  the age-sex structure of the population and the percentege
  of workers employed in the agriculture, forestry and fishing
  industries.}
\usage{data(scottish.lip.cancer)}
\format{ \tabular{rlll}{
    [,1] \tab observed \tab numeric \tab Number of oases \cr
    [,2] \tab expected \tab numeric  \tab expected number of cases \cr
    [,3] \tab AFF      \tab numeric  \tab percentage
    of workers employed in AFF \cr
  }
}
\source{Breslow, N.E. and Clayton, D.G,. (1993)
Approximate inference in generalized linear mixed models
\emph{J. Amer. Statistical Association}
pages 9-25, volume 88
}
\keyword{datasets}