\name{seeds}
\alias{seeds}
\title{Seeds germinating in a factorial experiment}
\description{
  This data set gives the number of seeds germinating in different
  plates of a factorial experiment, together with the total number of
  seeds in each plate, the seed variety and the root type}
\usage{data(seeds)}
\format{
  \tabular{rlll}{
    [,1] \tab r  \tab numeric  \tab Number of seeds germinating \cr
    [,2] \tab n  \tab numeric \tab Total number of seeds on plate \cr
    [,3] \tab x1 \tab numeric  \tab seed variety,
    (0 =  O. aegyptiaco 75,
    1 =  O. aegyptiaco 73) \cr
    [,4] \tab x2 \tab numeric \tab Root type (0 = bean, 1 = cucumber)}
}
\source{Crowder, M.J. (1978) Beta-Binomial ANOVA for Proportions
  \emph{Applied Statistics-J. Royal Statistical Soc. Series C}
  page 34-37, volume 27}
}
\keyword{datasets}